prettier and vitest config simplification
